+++ databasez/core/database.py
@@ -63,12 +63,6 @@ def init() -> None:
     )
 
 
-# we need a dict to ensure the references are kept
-ACTIVE_DATABASES: ContextVar[typing.Optional[typing.Dict[typing.Any, Database]]] = ContextVar(
-    "ACTIVE_DATABASES", default=None
-)
-
-
 ACTIVE_FORCE_ROLLBACKS: ContextVar[
     typing.Optional[weakref.WeakKeyDictionary[ForceRollback, bool]]
 ] = ContextVar("ACTIVE_FORCE_ROLLBACKS", default=None)
@@ -149,6 +143,7 @@ class Database:
     """
 
     _connection_map: weakref.WeakKeyDictionary[asyncio.Task, Connection]
+    _databases_map: typing.Dict[typing.Any, Database]
     _loop: typing.Any = None
     backend: interfaces.DatabaseBackend
     url: DatabaseURL
@@ -190,6 +185,7 @@ class Database:
         self._force_rollback = ForceRollback(force_rollback)
         self.backend.owner = self
         self._connection_map = weakref.WeakKeyDictionary()
+        self._databases_map = {}
 
         # When `force_rollback=True` is used, we use a single global
         # connection, within a transaction that always rolls back.
@@ -242,11 +238,18 @@ class Database:
     async def connect_hook(self) -> None:
         """Refcount protected connect hook. Executed begore engine and global connection setup."""
 
-    @multiloop_protector(True)
     async def connect(self) -> bool:
         """
         Establish the connection pool.
         """
+        loop = asyncio.get_running_loop()
+        if self._loop is not None and loop != self._loop:
+            # copy when not in map
+            if loop not in self._databases_map:
+                self._databases_map[loop] = self.__copy__()
+            # forward call
+            return await self._databases_map[loop].connect()
+
         if not await self.inc_refcount():
             assert self.is_connected, "ref_count < 0"
             return False
@@ -269,11 +272,14 @@ class Database:
     async def disconnect_hook(self) -> None:
         """Refcount protected disconnect hook. Executed after connection, engine cleanup."""
 
-    @multiloop_protector(True)
-    async def disconnect(self, force: bool = False) -> bool:
+    @multiloop_protector(True, inject_parent=True)
+    async def disconnect(
+        self, force: bool = False, *, parent_database: typing.Optional[Database] = None
+    ) -> bool:
         """
         Close all connections in the connection pool.
         """
+        # parent_database is injected and should not be specified manually
         if not await self.decr_refcount() or force:
             if not self.is_connected:
                 logger.debug("Already disconnected, skipping disconnection")
@@ -282,6 +288,14 @@ class Database:
                 logger.warning("Force disconnect, despite refcount not 0")
             else:
                 return False
+        if parent_database is not None:
+            loop = asyncio.get_running_loop()
+            del parent_database._databases_map[loop]
+        if force:
+            for sub_database in self._databases_map.values():
+                await sub_database.disconnect(True)
+            self._databases_map = {}
+        assert not self._databases_map, "sub databases still active"
 
         try:
             assert self._global_connection is not None
@@ -301,19 +315,10 @@ class Database:
         return True
 
     async def __aenter__(self) -> "Database":
+        await self.connect()
+        # get right database
         loop = asyncio.get_running_loop()
-        database = self
-        if self._loop is not None and loop != self._loop:
-            dbs = ACTIVE_DATABASES.get()
-            if dbs is None:
-                dbs = {}
-            else:
-                dbs = dbs.copy()
-            database = self.__copy__()
-            dbs[loop] = database
-            # it is always a copy required to prevent sideeffects between the contexts
-            ACTIVE_DATABASES.set(dbs)
-        await database.connect()
+        database = self._databases_map.get(loop, self)
         return database
 
     async def __aexit__(
@@ -322,13 +327,7 @@ class Database:
         exc_value: typing.Optional[BaseException] = None,
         traceback: typing.Optional[TracebackType] = None,
     ) -> None:
-        loop = asyncio.get_running_loop()
-        database = self
-        if self._loop is not None and loop != self._loop:
-            dbs = ACTIVE_DATABASES.get()
-            if dbs is not None:
-                database = dbs.pop(loop, database)
-        await database.disconnect()
+        await self.disconnect()
 
     @multiloop_protector(False)
     async def fetch_all(