+++ tests/test_databases.py
@@ -897,6 +897,13 @@ async def test_multi_thread_db_contextmanager(database_url):
             # many parallel and nested threads
             async with database as new_database:
                 await new_database.fetch_one("SELECT 1")
+                # test delegate to sub database
+                assert database.engine is new_database.engine
+                # also this shouldn't fail because redirected
+                old_refcount = new_database.ref_counter
+                await database.connect()
+                assert new_database.ref_counter == old_refcount + 1
+                await database.disconnect()
                 ops = []
                 while depth >= 0:
                     depth -= 1
@@ -909,14 +916,39 @@ async def test_multi_thread_db_contextmanager(database_url):
 
 
 @pytest.mark.asyncio
-async def test_multi_thread_db_connect_fails(database_url):
+async def test_multi_thread_db_connect(database_url):
     async with Database(database_url, force_rollback=True) as database:
 
         async def db_connect():
             await database.connect()
+            await database.fetch_one("SELECT 1")
+            await database.disconnect()
+
+        await to_thread(asyncio.run, db_connect())
+
+
+@pytest.mark.asyncio
+async def test_multi_thread_db_fails(database_url):
+    async with Database(database_url, force_rollback=True) as database:
+
+        async def db_connect():
+            # not in same loop
+            database.disconnect()
 
         with pytest.raises(RuntimeError):
             await to_thread(asyncio.run, db_connect())
+
+
+@pytest.mark.asyncio
+async def test_error_on_passed_parent_database(database_url):
+    database = Database(database_url)
+    # don't allow specifying parent_database
+    with pytest.raises(AssertionError):
+        await database.disconnect(parent_database=None)
+    with pytest.raises(AssertionError):
+        await database.disconnect(parent_database="")
+    with pytest.raises(TypeError):
+        await database.disconnect(False, None)
 
 
 @pytest.mark.asyncio